requirements.txt
attrs==23.1.0
branca==0.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
daftlistings==2.0.4
folium==0.15.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
